python/langchain-inference-masking/pyproject.toml
[tool.poetry]
name = "langchain-inference-masking"
version = "0.1.0"
description = "LangChain middleware example for PII masking during inference"
authors = ["John Sosoka"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^1.0"
langchain-openai = "^0.3"
langgraph = "^1.0"
pyyaml = "^6.0"
python-dotenv = "^1.2.1"
langgraph-prebuilt = "^1.0.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
ruff = "^0.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
